Move publications into markdown files.
Papers are grouped by year, with an additional special case of
"Preprints".  The structure of this grouping can be represented in YAML
as follows:

groupedPapers:
- year: Preprints
  papers:
  - a-preprint-paper
- year: 2017
  papers:
  - published-paper-1
  - published-paper-2
- year: 2016
  papers:
  - published-paper-3

This structure ensures ordering across the years, i.e. enables preprints
to come first, followed by each of the years in order, as well as also
preserving the order of papers within a year (i.e. paper-1 always
appears before paper-2 in 2017). A more naive approach of mapping each
array of papers to a year would not guarantee ordering of the years.
This approach would be represented in YAML as follows:

papers:
  Preprints:
  - a-preprint-paper
  2017:
  - published-paper-1
  - published-paper-2
  2016:
  - published-paper-3